fix: runpy run optimized fix: save environment cancel if target file doesn't exist

+++ nearai/environment.py
@@ -241,7 +241,10 @@ class Environment(object):
 
     def call_agent(self, agent_path: int, task: str) -> None:
         """Calls agent with given task."""
-        self._agents[agent_path].run(self, task=task)
+        try:
+            self._agents[agent_path].run(self, task=task)
+        finally:
+            print("end!!")
 
     def get_agents(self) -> List[Agent]:
         """Returns list of agents available in environment."""
@@ -285,47 +288,51 @@ class Environment(object):
         name = run_name or generated_name
 
         tempdir = Path(tempfile.mkdtemp())
+        environment_path = tempdir / "environment.tar.gz"
 
-        with open(tempdir / "environment.tar.gz", "r+b") as f:
-            with tarfile.open(fileobj=f, mode="w:gz") as tar:
-                tar.add(path, arcname=".")
-            f.flush()
-            f.seek(0)
-            snapshot = f.read()
-            tar_filename = f.name
+        if os.path.exists(environment_path):
+            with open(environment_path, "r+b") as f:
+                with tarfile.open(fileobj=f, mode="w:gz") as tar:
+                    tar.add(path, arcname=".")
+                f.flush()
+                f.seek(0)
+                snapshot = f.read()
+                tar_filename = f.name
 
-            timestamp = datetime.now(timezone.utc).isoformat()
+                timestamp = datetime.now(timezone.utc).isoformat()
 
-            entry_location = registry.upload(
-                tempdir,
-                EntryMetadata.from_dict(
-                    {
-                        "name": name,
-                        "version": "0.0.1",
-                        "description": f"Agent {run_type} run {agent_name}",
-                        "category": "environment",
-                        "tags": ["environment"],
-                        "details": {
-                            "base_id": base_id,
-                            "timestamp": timestamp,
-                            "agents": [agent.name for agent in self._agents],
-                            "run_id": run_id,
-                            "run_type": run_type,
-                            "filename": tar_filename,
-                        },
-                        "show_entry": True,
-                    }
-                ),
-                show_progress=True,
-            )
+                entry_location = registry.upload(
+                    tempdir,
+                    EntryMetadata.from_dict(
+                        {
+                            "name": name,
+                            "version": "0.0.1",
+                            "description": f"Agent {run_type} run {agent_name}",
+                            "category": "environment",
+                            "tags": ["environment"],
+                            "details": {
+                                "base_id": base_id,
+                                "timestamp": timestamp,
+                                "agents": [agent.name for agent in self._agents],
+                                "run_id": run_id,
+                                "run_type": run_type,
+                                "filename": tar_filename,
+                            },
+                            "show_entry": True,
+                        }
+                    ),
+                    show_progress=True,
+                )
 
-            location_str = plain_location(entry_location)
+                location_str = plain_location(entry_location)
 
-            print(f"Saved environment {entry_location} to registry. To load use flag `--load-env={location_str}`.")
+                print(f"Saved environment {entry_location} to registry. To load use flag `--load-env={location_str}`.")
 
-        rmtree(tempdir)
-
-        return snapshot
+            rmtree(tempdir)
+            return snapshot
+        else:
+            print(f"The file {environment_path} does not exist.")
+            return None
 
     def load_snapshot(self, snapshot: bytes) -> None:
         """Load Environment from Snapshot."""
@@ -361,6 +368,10 @@ class Environment(object):
     def request_user_input(self) -> None:
         """Must be called to request input from the user."""
         self.set_next_actor("user")
+
+    def clear_temp_agent_files(self) -> None:  # noqa: D102
+        """Remove temp agent files created to be used in `runpy`."""
+        shutil.rmtree(self._agents[0].temp_dir)
 
     def set_next_actor(self, who: str) -> None:  # noqa: D102
         next_action_fn = os.path.join(self._path, ".next_action")
@@ -415,6 +426,10 @@ class Environment(object):
                 self.add_message("user", new_message)
 
                 self.set_next_actor("agent")
+
+        self.clear_temp_agent_files()
+
+        print("record_run", record_run)
 
         if record_run:
             run_name = record_run if record_run and record_run != "true" else None
